Move AVX-512 availability testing code to rten-vecmath
This is a prerequisite to using AVX-512 more widely, outside of the GEMM code.
